update thair branch

+++ src/main/java/model/Member.java
@@ -17,7 +17,7 @@ public class Member{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	int memberId;
+	Long memberId;
 	
 	String firstName;
 	String lastName;
@@ -61,12 +61,12 @@ public class Member{
 
 
 
-	public int getMemberId() {
+	public Long getMemberId() {
 		return memberId;
 	}
 
 
-	public void setMemberId(int memberId) {
+	public void setMemberId(Long memberId) {
 		this.memberId = memberId;
 	}
